Collection extension methods all(), none(), all() moved to newly created Sequence Extension
SequenceExtensions.swift file added
SequenceExtensionsTests.swift file added
