+++ Tests/SwiftStdlibTests/CollectionExtensionsTests.swift
@@ -30,19 +30,4 @@ final class CollectionExtensionsTests: XCTestCase {
         XCTAssert([1, 2, 3].contains([1, 2, 3].randomItem!))
         XCTAssertNil([].randomItem)
     }
-    
-    func testAllMatch() {
-        let collection = [2, 4, 6, 8, 10, 12]
-        XCTAssert(collection.all { $0 % 2 == 0 })
-    }
-    
-    func testAnyMatch() {
-        let collection = [3, 5, 8, 9, 11, 13]
-        XCTAssert(collection.any { $0 % 2 == 0 })
-    }
-    
-    func testNoneMatch() {
-        let collection = [3, 5, 7, 9, 11, 13]
-        XCTAssert(collection.none { $0 % 2 == 0 })
-    }
 }
